+++ src/QafooLabs/Bundle/NoFrameworkBundle/Tests/Controller/ResultConverter/HeadersYieldApplierTest.php
@@ -12,7 +12,7 @@ class HeadersYieldApplierTest extends TestCase
 {
     private $applier;
 
-    public function setUp()
+    public function setUp() : void
     {
         $this->applier = new HeadersYieldApplier();
     }